+++ wserver/src/main/java/com/wanderdrop/wserver/dto/AuthenticationResponse.java
@@ -1,9 +1,12 @@
-package com.wanderdrop.wserver.auth;
+package com.wanderdrop.wserver.dto;
 
+import com.wanderdrop.wserver.model.Role;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.util.UUID;
 
 @Data
 @Builder
@@ -11,4 +14,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class AuthenticationResponse {
     private String token;
+
+    private Role role;
+
+    private UUID userId;
 }
